+++ app/src/main/java/com/bt/zhangzy/network/entity/JsonCarTrack.java
@@ -10,7 +10,7 @@ import java.util.Date;
 public class JsonCarTrack extends BaseEntity {
 
     int id, driverId, orderId, carId;
-    float longitude, latitude;
+    double longitude, latitude;
     Date createDate;
 
     public int getId() {
@@ -45,19 +45,19 @@ public class JsonCarTrack extends BaseEntity {
         this.carId = carId;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
